move : api 패키지 domain > common 으로 이동
