+++ src/main/java/com/gachon/home_protector/domain/common/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.gachon.home_protector.api;
+package com.gachon.home_protector.domain.common;
 
 import lombok.Builder;
 import lombok.Getter;
