+++ src/main/java/com/gachon/home_protector/security/handler/RestAuthenticationSuccessHandler.java
@@ -1,7 +1,7 @@
 package com.gachon.home_protector.security.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.gachon.home_protector.api.SuccessResponse;
+import com.gachon.home_protector.domain.common.SuccessResponse;
 import com.gachon.home_protector.token.token.RefreshToken;
 import com.gachon.home_protector.token.repository.RefreshTokenRepository;
 import com.gachon.home_protector.security.userdetails.RestUserDetails;
