+++ src/main/java/com/gachon/home_protector/user/User.java
@@ -1,6 +1,6 @@
 package com.gachon.home_protector.user;
 
-import com.gachon.home_protector.api.BaseEntity;
+import com.gachon.home_protector.domain.common.BaseEntity;
 import com.gachon.home_protector.user.dto.join.RestUserJoinResponse;
 import com.gachon.home_protector.user.dto.login.RestUserLoginResponse;
 import jakarta.persistence.*;
