+++ src/main/java/com/gachon/home_protector/user/UserController.java
@@ -1,6 +1,6 @@
 package com.gachon.home_protector.user;
 
-import com.gachon.home_protector.api.SuccessResponse;
+import com.gachon.home_protector.domain.common.SuccessResponse;
 import com.gachon.home_protector.security.userdetails.RestUserDetails;
 import com.gachon.home_protector.user.annotation.valid_identification_token.ValidateIdentificationToken;
 import com.gachon.home_protector.user.dto.identification.IdentificationRequest;
